Finished the python scripts for creating massive numbers of accounts and massive numbers of submissions for testing purposes. Found a few shaky spots in the compile pipeline and the tournament manager and fixed those as well.
git-svn-id: https://ai-contest.googlecode.com/svn/trunk@153 9d0e4ec5-6d14-e29b-3149-cabf4038ef1e

+++ planet_wars/integration_testing/delete_test_accounts.py
@@ -7,6 +7,13 @@ connection = MySQLdb.connect(host = server_info["db_host"],
                              passwd = server_info["db_password"],
                              db = server_info["db_name"])
 cursor = connection.cursor(MySQLdb.cursors.DictCursor)
-cursor.execute("DELETE FROM users WHERE username LIKE 'testbot%'")
+query = "SELECT user_id, username FROM users WHERE username LIKE 'testbot%'"
+cursor.execute(query)
+accounts = cursor.fetchall()
+if len(accounts) > 0:
+  submission_ids = ",".join([str(acc["user_id"]) for acc in accounts])
+  query = "DELETE FROM submissions WHERE user_id IN (" + submission_ids + ")"
+  cursor.execute(query)
+  cursor.execute("DELETE FROM users WHERE username LIKE 'testbot%'")
 cursor.close()
 connection.close()
